requirements.txt
PyMySQL==1.0.2
prompt_toolkit==3.0.14
python-dotenv==0.15.0
rich==9.10.0
cryptography==3.3.1
